fix: apply PR feedback
