+++ package.json
@@ -4,11 +4,10 @@
   "private": true,
   "packageManager": "pnpm@9.12.0",
   "devDependencies": {
-    "@biomejs/biome": "^1.9.4",
-    "@types/node": "^20.9.0",
-    "typescript": "^5.2.2",
-    "tsup": "^8.3.5",
-    "vitest": "^3.0.3"
+    "@biomejs/biome": "^1.9.4"
+  },
+  "engines": {
+    "node": ">=18.14"
   },
   "resolutions": {
     "vite": "5.4.12"
