+++ packages/ensnode-common/package.json
@@ -3,24 +3,25 @@
   "version": "0.0.1",
   "type": "module",
   "exports": {
-    ".": "./src/index.ts",
     "./helpers": "./src/helpers.ts",
     "./ids": "./src/ids.ts",
     "./plugin-helpers": "./src/plugin-helpers.ts",
     "./subname-helpers": "./src/subname-helpers.ts"
   },
   "scripts": {
+    "lint": "biome check --write .",
+    "lint:ci": "biome ci",
     "test": "vitest",
     "typecheck": "tsc --noEmit"
   },
-  "dependencies": {
-    "ts-deepmerge": "^7.0.2",
+  "peerDependencies": {
     "viem": "^2.21.57"
   },
   "devDependencies": {
     "@biomejs/biome": "^1.9.4",
     "@types/node": "^20.9.0",
     "typescript": "^5.2.2",
+    "viem": "^2.21.57",
     "vitest": "^3.0.3"
   },
   "engines": {
