+++ packages/ensnode-subgraph-graphql/README.md
@@ -1,0 +1,3 @@
+# ENSNode Subgraph GraphQL
+
+ENSNode Subgraph GraphQL is a Hono middleware for generating Subgraph-compatible GraphQL schema.
