+++ packages/ensnode-subgraph-graphql/package.json
@@ -3,7 +3,6 @@
   "version": "0.0.1",
   "type": "module",
   "exports": {
-    ".": "./src/index.ts",
     "./middleware": "./src/middleware.ts"
   },
   "scripts": {
